requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
kavenegar==1.1.2
pillow==11.1.0
PyJWT==2.9.0
redis==5.2.1
requests==2.32.3
smsir-python==1.0.8
sqlparse==0.5.3
urllib3==2.3.0
